requirements.txt
anyio==3.6.2
bidict==0.21.2
click==8.0.1
fastapi==0.88.0
fastapi-socketio==0.0.9
Flask==2.0.1
Flask-SocketIO==5.1.1
h11==0.14.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.1
libvirt-python==8.10.0
MarkupSafe==2.0.1
pip-autoremove==0.10.0
pkgconfig==1.5.5
pydantic==1.10.2
python-dotenv==0.21.0
python-engineio==4.2.1
python-socketio==5.4.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.0.1
